requirements.txt
# Backend Core
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2.5
python-multipart>=0.0.6

# Search & Embeddings (CPU-only)
rank-bm25>=0.2
sentence-transformers>=2.3
faiss-cpu>=1.8
numpy>=1.24
torch>=2.2

# Database & Storage
sqlalchemy>=2.0
aiosqlite>=0.19

# Utilities
python-dotenv>=1.0
tqdm>=4.66

# Testing
pytest>=7.4
pytest-asyncio>=0.23
httpx>=0.26

# Logging & Metrics
structlog>=24.1